Move producers to src directory
