--- HLTrigger/Muon/plugins/BuildFile.xml
@@ -1,7 +1,0 @@
-<library   file="*.cc" name="HLTriggerMuonPlugins">
-  <use   name="DataFormats/Common"/>
-  <use   name="DataFormats/Scouting"/>
-  <use   name="FWCore/Framework"/>
-  <use   name="DataFormats/TrackReco"/>
-  <flags   EDM_PLUGIN="1"/>
-</library>
